refactor: Api cleaning
